Testing main path and some minor fixing.
