+++ src/Stratis.Bitcoin/Consensus/ConsensusExceptions.cs
@@ -1,0 +1,36 @@
+﻿using System;
+
+namespace Stratis.Bitcoin.Consensus
+{
+    public class ConsensusException : Exception
+    {
+        public ConsensusException() : base()
+        {
+        }
+
+        public ConsensusException(string messsage) : base(messsage)
+        {
+        }
+    }
+
+    public class MaxReorgViolationException : ConsensusException
+    {
+        public MaxReorgViolationException() : base()
+        {
+        }
+    }
+
+    public class ConnectHeaderException : ConsensusException
+    {
+        public ConnectHeaderException() : base()
+        {
+        }
+    }
+
+    public class InvalidHeaderException : ConsensusException
+    {
+        public InvalidHeaderException() : base()
+        {
+        }
+    }
+}
